feat: remove `CompletionQueue` from InstanceAdmin
This is a part of #2567.

This PR removes the `CompletionQueue` parameter from the `InstanceAdmin`
API. Apart from the obvious changes, an additional option to
`bigtable::ConnectionOptions` was added to allow for custom worker
threads.

A decision was made to store the background threads in `InstanceAdmin`
rather than `CommonClient` even though `CommonClient` seems to resemble
other clients better. This is because `CommonClient` must not be the
owner of the background threads executing the completion queue
associated with itself because at some point the completion queue may hold
the last reference to the `CommonClient`. This is problematic because it
could potentially lead to a situation where `CompletionQueue` would
trigger `CommonClient`'s destructor, which in turn would trigger the
same `CompletionQueue`'s destructor, which would be really hard to
handle.

This is a reiteration of #5665.